requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
face_recognition==1.3.0
numpy==1.26.4
Pillow==10.3.0
